requirements.txt
uvicorn
pymongo
bcrypt
python-dotenv
python-multipart
jinja2==3.1.4
starlette==0.37.2
fastapi==0.115.0
itsdangerous>=2.1.2